leo/requirements.txt
polygon-api-client>=1.14.0
mysql-connector-python>=8.0.33
pandas>=2.0.0
numpy>=1.26.0
flask>=3.0.0
alpaca-py>=0.20.0
